Reverting the switch/break modification I made to Num::ordinal() the other day.
It turns out there was a bug with sniff for indentation of switch cases,
at the time I wrongly assumed it was due to the return statement in the
cases.  Said bug has been fixed and there's a now a testcase for it.

Apologies for all the controversy this caused, I should probably stop
committing code at 3AM