CSRF provider should return dummy test data if headers have been sent
For some strange reason it seems providers run on all tests in your testsuite,
even if your excluding them via groups. As the CSRF provider uses the session
class it errors out if headers have already been sent.

This implementation is a bad hack, the security class should really be re-written
to allow a session instance to be injected

fixes #4155