Remove physical path dependency in Kohana::module tests
Backports dd926956 to 3.3 - MODPATH is unreliable in a koharness
environment as it will usually be a symlink. With this change the test
of Kohana::modules will pass wherever the code happens to be
deployed on the disk. This will conflict when we merge up to 3.4
because that commit uses short array syntax which is not available
in the 3.3 series.