Reinstate one skipped test for File::mime
These tests aren't great, and the PHP mime-type test is particularly likely to be
unreliable. We should be able to get a valid (eg image/png) result in all current
environments though for the moment.

Probably the File::mime method would be best reimplemented with some
third-party component, surely this is a solved problem by now.