Reinstate and improve Feed::parse tests
Original test was being skipped because it was hitting the internet
- the major relevant functionality here is the parsing rather than
the remote access.

Added local example feeds for RSS2.0 as well as the previously
tested atom feed. Also improved the tests to check for the title
as well as the number of elements since we now have known
content to work with.