Remove invalid skipped test - Response does send headers in CLI
This test was originally implemented in bf0175f6cd to fix
http://dev.kohanaframework.org/issues/3767 - it was required
because headers were sent directly by the response class, and
once they had been sent it was then very difficult to test other
variations. The fact that headers are not relevant in CLI appears
to have been a side concern.

Subsequently in a1943b0d7d26cf5a header sending moved to
HTTP_Header which mocks out `header` with a callback interface
for testing instead.

Since that refactoring (released in v3.2.0) the headers are always
sent in CLI unless headers_sent is already true.

Therefore this test is invalid and can be removed.