Declare required logger interfaces
Adds a Kohana_Logger interface that basically extends the PRS-3
LoggerInterface, with additional methods for deffered logging.

At this point any implementation of Kohana_Logger interface is
tied to the Log_Writer class. We can later decouple it by
introducing a separate interface to abstract the Log_Writer.

Also adds a Logger interface by basically extends Kohana_Logger.
This is just for the CFS enhanced extension, to keep the trend
with other interfaces in Kohana. IMHO, however, I can not
understand the value for this, as basically, a tweeked interface
can make more harm than any benefit.