Remove *trace*, *class*, and *function* from log message
Removed *trace*, *class*, and *function* from log message, because we
can not have the "same log message results" when logs are generated from
*level* specific methods (`emergency`, `debug`, ...) vs the generic
`log` method.

It is worthy to note that the above elements are not used by the
`abstract class Kohana_Log_Writer`, so they are not usually logged,
unless Kohana_Log_Writer is extended by userland, using CFS.

Added a fuzzy `assertLogMessageEquals` to the test class to make log
message testing easier, because generating log messages always include
the line number from which the `log` (or equivalent *level* methods) are
called.