Restructure Request tests to avoid Request::factory calls in providers
`Request::factory()` calls in test "provider" methods initializes the
`Request::$initial` static variable, affecting other test methods, even
if those "provider" methods are not direct providers of those other test
methods.

In our case the `HTTPTest::test_redirect` is affected because the
`HTTP::redirect` method needs the `Request::$initial` static variable to
be available.

Even if PHPUnit is run with the `--filter=HTTPTest` flag, the "provider"
methods of `RequestTest` are still available, `Request::factory()` calls
are still made, and the `Request::$initial` static variable is still
initialized.

This becomes problematic because `HTTPTest::test_redirect` is now
dependant of the `RequestTest` "providers", as those "providers" change
the global state in favor of its needs. Once `RequestTest` "providers"
are not available -- ex: by deleting RequestTest.php file from disk --
`HTTPTest::test_redirect` test cases fail.

In order to not alter the global state, and to not affect other tests,
static variables initialization, and calls to methods that affect static
variables should not be made in test case providers. Rather calls
should be done carefully in `setUp` and `tearDown` methods, as well as
in the tests methods themselves.

In this case I have removed `Request::factory()` calls from providers
and restructed the tests.